+++ path-tracker-pipelines/src/main/java/org/example/Main.java
@@ -31,6 +31,7 @@ import java.util.Collections;
 import java.util.Properties;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class Main {
 
@@ -88,7 +89,7 @@ public class Main {
         }
 
 
-        AtomicInteger watermarks = new AtomicInteger();
+        AtomicLong watermarks = new AtomicLong();
 
         // Make producer, consumer, and merger
         KafkaMergeThread mergeThread = new KafkaMergeThread(pathNum, queue, watermarks);
